@@ -4,33 +4,34 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Quote {
-	private String type;
-	private Value value;
-	
-	public Quote(){}
+    private String type;
+    private Value value;
 
-	public String getType() {
-		return type;
-	}
+    public Quote() {
+    }
 
-	public void setType(String type) {
-		this.type = type;
-	}
+    public String getType() {
+        return type;
+    }
 
-	public Value getValue() {
-		return value;
-	}
+    public void setType(String type) {
+        this.type = type;
+    }
 
-	public void setValue(Value value) {
-		this.value = value;
-	}
-	
-	@Override
-	public String toString() {
-		return "Quote{" +
-		        "type='" + type + '\'' +
-		        ", value=" + value +
-		        '}';
-	}
+    public Value getValue() {
+        return value;
+    }
+
+    public void setValue(Value value) {
+        this.value = value;
+    }
+
+    @Override
+    public String toString() {
+        return "Quote{" +
+                "type='" + type + '\'' +
+                ", value=" + value +
+                '}';
+    }
 
 }
